Custom requests UI changes Route changes

+++ database/migrations/2016_12_05_092743_create_routes_table.php
@@ -17,7 +17,7 @@ class CreateRoutesTable extends Migration
             $table->increments('id');
             $table->string('user_id')->references('id')->on('users');
             $table->string('car_id')->references('id')->on('cars');
-            $table->integer('distance_travelled');
+            $table->double('distance_travelled');
             $table->double('total_cost');
             $table->timestamps();
         });
